third_party/blink/web_tests/external/wpt/websockets/interfaces/WebSocket/events/016_wpt_flags=h2-expected.txt

This is a testharness.js-based test.
FAIL WebSockets: addEventListener assert_equals: expected 7 but got 1
Harness: the test ran to completion.
